feat: wire end-to-end generation pipeline and deploy backend
- Connect iOS pipeline: Canvas → Preprocessor → APIClient → ResultView
- Add generate button (sparkles) to floating toolbar
- AppCoordinator orchestrates snapshot capture, preprocessing, and API call
- Implement latest-request-wins with staleness checks
- ResultView shows shimmer during generation, preserves last image on error
- Fix fal.ai adapter: switch to fast-sdxl/image-to-image (LCM model broken)
- Use synchronous fal.run endpoint instead of queue polling
- Fix seed overflow: use UInt64 for large fal.ai seed values
- Fix unused orchestratorPlugin import blocking production build
- Render PencilKit snapshots on white background before sending
- Deploy backend to Railway (kiki-backend-production-eb81.up.railway.app)
- Add sharp dependency for image processing
- Tune generation params (strength, guidance_scale) for sketch-to-image

+++ ios/Packages/NetworkModule/Sources/NetworkModule/GenerateResponse.swift
@@ -4,7 +4,7 @@ public struct GenerateResponse: Sendable {
     public let requestId: UUID
     public let status: ResponseStatus
     public let imageURL: URL?
-    public let seed: Int?
+    public let seed: UInt64?
     public let provider: String?
     public let latencyMs: Int?
     public let mode: GenerationMode
@@ -13,7 +13,7 @@ public struct GenerateResponse: Sendable {
         requestId: UUID,
         status: ResponseStatus,
         imageURL: URL? = nil,
-        seed: Int? = nil,
+        seed: UInt64? = nil,
         provider: String? = nil,
         latencyMs: Int? = nil,
         mode: GenerationMode
